Renamed app to match dev branch.
